+++ server/tests/route.spec.js
@@ -1,6 +1,6 @@
 import chai from 'chai';
 import chaiHttp from 'chai-http';
-import app from '../routes/index.js';
+import app from '../routes/Index.js';
 
 process.env.NODE_ENV = 'test';
